Modify the DNS deployment
Added a workaround on how the additional DNS Servers are added to a zone
file. Now it is possible to change the list based on a variable and thus
use an empty list.

The additional dns servers are not really a secret which is why they are
listet in the group_vars instead of the secret files

close #2